move styles for NavBar.js into NarBarStyles.js
