+++ frontend/src/components/NavBar/NavBar.js
@@ -1,6 +1,7 @@
 import React from "react";
+import { Menu, ChevronLeft, ChevronRight } from "@material-ui/icons";
+import { useStyles, useTheme } from './NavBarStyles';
 import clsx from "clsx";
-import { makeStyles, useTheme } from "@material-ui/core/styles";
 import {
   Drawer,
   CssBaseline,
@@ -14,90 +15,10 @@ import {
   ListItemText,
   Button,
 } from "@material-ui/core";
-import { Menu, ChevronLeft, ChevronRight } from "@material-ui/icons";
 
-const drawerWidth = 250;
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    display: "flex",
-    backgroundColor: "#F0F5EF",
-    color: "#000000",
-  },
-  title: {
-    flexGrow: 1,
-  },
-  close_btn: {
-    [theme.breakpoints.up("xs")]: {
-      display: "none",
-    },
-    [theme.breakpoints.up("sm")]: {
-      display: "flex",
-    },
-  },
-  appBar: {
-    transition: theme.transitions.create(["margin", "width"], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-  },
-  appBarShift: {
-    width: `calc(100% - ${drawerWidth}px)`,
-    marginLeft: drawerWidth,
-    transition: theme.transitions.create(["margin", "width"], {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-    [theme.breakpoints.up("sm")]: {
-      display: "none",
-    },
-  },
-  hide: {
-    display: "none",
-  },
-  drawer: {
-    width: drawerWidth,
-    flexShrink: 0,
-    [theme.breakpoints.up("sm")]: {
-      width: drawerWidth,
-      flexShrink: 0,
-    },
-  },
-  drawerPaper: {
-    width: drawerWidth,
-  },
-  drawerHeader: {
-    display: "flex",
-    alignItems: "center",
-    padding: theme.spacing(0, 1),
-    // necessary for content to be below app bar
-    ...theme.mixins.toolbar,
-    justifyContent: "flex-end",
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(3),
-    transition: theme.transitions.create("margin", {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-    marginLeft: -drawerWidth,
-  },
-  contentShift: {
-    transition: theme.transitions.create("margin", {
-      easing: theme.transitions.easing.easeOut,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-    marginLeft: 0,
-  },
-}));
-
-export default function PersistentDrawerLeft() {
-  const classes = useStyles();
-  const theme = useTheme();
+export default function PersistentDrawerLeft(props) {
+  const classes = useStyles(props);
+  const theme = useTheme(props);
   const [open, setOpen] = React.useState(false);
 
   const handleDrawerOpen = () => {
